requirements.txt
# NOTE: I experimented with text generation in this virtualenv, but did not use it.
# Some packages below will not be needed to run final scripts.
# https://www.tensorflow.org/text/tutorials/text_generation

absl-py==1.2.0
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bert-extractive-summarizer==0.10.1
bertviz==1.4.0
bleach==5.0.1
blis==0.7.8
boto3==1.24.42
botocore==1.27.42
bpemb==0.3.3
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
commonmark==0.9.1
conllu==4.5.2
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
filelock==3.7.1
flair==0.11.3
flatbuffers==1.12
fonttools==4.34.4
ftfy==6.1.1
future==0.18.2
gast==0.4.0
gdown==4.4.0
gensim==4.2.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==3.7.0
huggingface-hub==0.8.1
hyperopt==0.2.7
idna==3.3
importlib-metadata==3.10.1
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
Janome==0.4.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.8
jsonschema==4.9.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keybert==0.6.0
kiwisolver==1.4.4
konoha==4.6.5
langcodes==3.3.0
langdetect==1.0.9
libclang==14.0.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.13.0
mpld3==0.3
murmurhash==1.0.7
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
pptree==3.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.9.0
segtok==1.5.11
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.95
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
sqlitedict==2.0.0
srsly==2.4.4
stack-data==0.3.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
thinc==8.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.0
typer==0.4.2
typing-extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.11
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.1
widgetsnbextension==3.6.1
Wikipedia-API==0.5.4
wordcloud==1.8.2.2
wrapt==1.14.1
zipp==3.8.1
